+++ CHANGES.md
@@ -1,5 +1,12 @@
 twilio-php Changelog
 ====================
+[2023-02-22] Version 7.0.0-rc.4
+-------------------------------
+**Api**
+- Remove `scheduled_for` property from message resource
+- Add `scheduled_for` property to message resource
+
+
 [2023-02-08] Version 7.0.0-rc.3
 -------------------------------
 **Library - Fix**